Repackage by feature instead of layer
