+++ src/main/java/org/commonwl/viewer/cwl/CWLElement.java
@@ -17,31 +17,24 @@
  * under the License.
  */
 
-package org.commonwl.viewer.domain.cwl;
+package org.commonwl.viewer.cwl;
 
-import java.util.Map;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
- * Represents a step of a workflow
+ * Represents the input/output of a workflow/tool
  */
-public class CWLStep {
+public class CWLElement {
 
     private String label;
     private String doc;
     private String type;
-    private String run;
-    private CWLProcess runType;
-    private Map<String, CWLElement> inputs;
-    private Map<String, CWLElement> outputs;
+    private List<String> sourceID;
+    private String defaultVal;
 
-    public CWLStep(String label, String doc, String type, String run,
-                   Map<String, CWLElement> inputs, Map<String, CWLElement> outputs) {
-        this.label = label;
-        this.doc = doc;
-        this.type = type;
-        this.run = run;
-        this.inputs = inputs;
-        this.outputs = outputs;
+    public CWLElement() {
+        this.sourceID = new ArrayList<String>();
     }
 
     public String getLabel() {
@@ -68,36 +61,21 @@ public class CWLStep {
         this.type = type;
     }
 
-    public String getRun() {
-        return run;
+    public List<String> getSourceIDs() {
+        return sourceID;
     }
 
-    public void setRun(String run) {
-        this.run = run;
+    public void addSourceID(String sourceID) {
+        if (sourceID != null) {
+            this.sourceID.add(sourceID);
+        }
     }
 
-    public CWLProcess getRunType() {
-        return runType;
+    public String getDefaultVal() {
+        return defaultVal;
     }
 
-    public void setRunType(CWLProcess runType) {
-        this.runType = runType;
+    public void setDefaultVal(String defaultVal) {
+        this.defaultVal = defaultVal;
     }
-
-    public Map<String, CWLElement> getInputs() {
-        return inputs;
-    }
-
-    public void setInputs(Map<String, CWLElement> inputs) {
-        this.inputs = inputs;
-    }
-
-    public Map<String, CWLElement> getOutputs() {
-        return outputs;
-    }
-
-    public void setOutputs(Map<String, CWLElement> outputs) {
-        this.outputs = outputs;
-    }
-
 }
