+++ src/main/java/org/commonwl/viewer/docker/DockerService.java
@@ -1,4 +1,4 @@
-package org.commonwl.viewer.services;
+package org.commonwl.viewer.docker;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
