+++ src/main/java/org/commonwl/viewer/github/GithubDetails.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.commonwl.viewer.domain;
+package org.commonwl.viewer.github;
 
 import java.io.Serializable;
