+++ src/main/java/org/commonwl/viewer/researchobject/HashableAgent.java
@@ -1,4 +1,4 @@
-package org.commonwl.viewer.domain.researchobject;
+package org.commonwl.viewer.researchobject;
 
 import org.apache.taverna.robundle.manifest.Agent;
