+++ src/main/java/org/commonwl/viewer/workflow/WorkflowNotFoundException.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.commonwl.viewer.web;
+package org.commonwl.viewer.workflow;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
